beacon/goclient: move beacon options from protocol/v2/blockchain/beacon to beacon/goclient
