+++ beacon/goclient/events_test.go
@@ -73,7 +73,7 @@ func TestSubscribeToHeadEvents(t *testing.T) {
 }
 
 func eventsTestClient(t *testing.T, serverURL string) *GoClient {
-	server, err := New(zap.NewNop(), beacon.Options{
+	server, err := New(zap.NewNop(), Options{
 		BeaconNodeAddr: serverURL,
 		Context:        context.Background(),
 		Network:        beacon.NewNetwork(types.MainNetwork),
